+++ lnbits/core/services/__init__.py
@@ -3,7 +3,7 @@ from .funding_source import (
     switch_to_voidwallet,
 )
 from .lnurl import perform_lnurlauth, redeem_lnurl_withdraw
-from .notifications import enqueue_notification
+from .notifications import enqueue_notification, send_payment_notification
 from .payments import (
     calculate_fiat_amounts,
     check_transaction_status,
@@ -39,6 +39,7 @@ __all__ = [
     "perform_lnurlauth",
     # notifications
     "enqueue_notification",
+    "send_payment_notification",
     # payments
     "calculate_fiat_amounts",
     "check_transaction_status",
